requirements.txt
Flask==3.1.3
Flask-Bootstrap
pandas
Flask-SQLAlchemy
pymysql
python-dotenv